添加了【門店自提】

+++ twant/app/src/main/java/com/ftofs/twant/entity/ChatMessage.java
@@ -7,7 +7,7 @@ public class ChatMessage {
     public static final int YOUR_MESSAGE = 2;
 
     public int origin; // 我發的消息還是別人發的消息
-    public int timestamp;
+    public long timestamp; // 毫秒
     public EMMessage.Type messageType;
     public String memberName;
     public String nickname;
